google_cloud_function/online_retail/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
requests==2.21.0
setuptools==40.2.0
Werkzeug==0.14.1
wheel==0.31.1
pandas==1.0.0
pyarrow	
gcsfs
google-cloud-storage
google-cloud-bigquery==1.24.0
